level-9 + partial gui fix
