+++ src/main/java/Task.java
@@ -1,4 +1,4 @@
-public class Task {
+abstract public class Task {
     protected String description;
     protected boolean isDone;
     protected String symbol;
